Update of multiple selection work for Form 3.2 completed. 1. Proposed Water Use 2. River Nature 3. Ground Water Quality 4. Water Diversion Source

+++ WrpCcNocWeb/Models/CcModule/CcModHydraInfraParamDetail.cs
@@ -23,7 +23,7 @@ namespace WrpCcNocWeb.Models
         [Required]
         [Column("HydraInfraParamId", Order = 2)]
         [Display(Name = "Parameter Name")]
-        public int HydraInfraParamId { get; set; }
+        public long HydraInfraParamId { get; set; }
         [ForeignKey("HydraInfraParamId")]
         public virtual LookUpCcModHydraInfraParam LookUpCcModHydraInfraParam { get; set; }
